Remove objc tests
